OF has been changed; 0.8 version isn't compatible with VS 2010, switched to 0.7.4

+++ OF/examples/graphics/lutFilterExample/src/testApp.cpp
@@ -128,7 +128,7 @@ void testApp::keyReleased(int key){
 			break;
 		case OF_KEY_UP:
 			dirLoadIndex++;
-			if (dirLoadIndex>=(int)dir.size()) {
+			if (dirLoadIndex>=dir.size()) {
 				dirLoadIndex=0;
 			}
 			loadLUT(dir.getPath(dirLoadIndex));
